requirements.txt
typer[all]
google-cloud-aiplatform
